+++ src/main/java/com/calcoa/pricing/model/Service.java
@@ -4,7 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name="services")
@@ -17,7 +21,7 @@ public class Service {
     @Column(name="name")
     public String name;
     @Column(name="price")
-    public Float price;
+    public BigDecimal price;
     @Column(name="type")
     public String type;
 }
